LibraryManageSystem/requirements.txt
# Flask Core
Flask==3.0.0
Flask-RESTX==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL adapter for production

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0

# HTTP Client
requests==2.31.0

# Rate Limiting
Flask-Limiter==3.5.0
redis==5.0.1  # For distributed rate limiting

# Monitoring & Logging
psutil==5.9.6  # System metrics
prometheus-client==0.19.0  # Prometheus metrics exporter
prometheus-flask-exporter==0.23.0  # Flask integration

# Production WSGI Server
gunicorn==21.2.0

# Development Tools
pytest==7.4.3
pytest-flask==1.3.0
